refactoring phantom movie generation: decreasing memory requirement speeding up processing generating movies with noise

+++ src/xmipp/applications/programs/phantom_movie/phantom_movie_main.cpp
@@ -26,4 +26,4 @@
 #include <core/xmipp_program.h>
 #include <reconstruction/phantom_movie.h>
 
-RUN_XMIPP_PROGRAM(PhantomMovie<float>)
+RUN_XMIPP_PROGRAM(PhantomMovie<double>)
